// apps/mobile/api/generated/arenaAPI.schemas.ts
export interface InviteCodeDto {
  /** 초대 코드 ID */
  inviteCodeId?: string;
  /** 초대 코드 */
  code?: string;
  /** 생성일시 */
  createdAt?: string;
}